correct errors in qt and catalysis
